pyproject.toml
[project]
name = "pyscripts"
version = "0.1.0"
description = ""
authors = [
    {name = "Vincent Ramdhanie", email = "vramdhanie@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "moviepy==1.0.3",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
    "google-api-python-client>=2.149.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
]

[tool.poetry]
name = "pyscripts"
version = "0.1.0"
description = ""
authors = ["Vincent Ramdhanie <vramdhanie@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
moviepy = "1.0.3"
pandas = "^2.0.0"
matplotlib = "^3.7.0"
numpy = "^1.24.0"
google-api-python-client = "^2.149.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
